+++ CHANGELOG.md
@@ -28,6 +28,7 @@
   * `find_by_player_id` to get all the rewards for a player by the player_id (reads through pagination)
   * `find_by_email_and_site` to get all the rewards for a player by email and site (reads through pagination)
   * `type` to get the type of the reward (e.g. level, achievement etc.)
+  * `definition` to get the reward definition if there was one
 * Badgeville::RewardDefinition implemented with:
   * `find_by_site` to get all the reward definitions for a site (returned as Badgeville::Reward)
 * Badgeville::Client now raises correct Errors when RestClient raises an error